refactor(sdk)!: move Api/Http/Auth types to iii.types in the Python SDK
BREAKING CHANGE: ApiRequest, ApiResponse, HttpRequest, HttpResponse, AuthInput and
AuthResult are no longer exported from the iii package root. Import them from iii.types.
